api/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pgvector
python-jose[cryptography]
passlib
bcrypt==3.2.0
python-multipart
email-validator

